Correction on idt_set_gate, the index-> num, must be numeric, we change to int
Added some basic declaration-only headers from libc

+++ kernel/cpu/idt.c
@@ -63,7 +63,7 @@ static void init_idt_table()
 	idt_flush((uint32_t)&idt_ptr);
 }
 
-void idt_set_gate(uint8_t num, uint32_t base, uint16_t sel, uint8_t flags)
+void idt_set_gate(int num, uint32_t base, uint16_t sel, uint8_t flags)
 {
 	idt_entries[num].base_lo = base & 0xFFFF;
 	idt_entries[num].base_hi = (base >> 16) & 0xFFFF;
